requirements.txt
langchain>=0.3.0
langchain-openai
langchain-community 
langchain-text-splitters
faiss-cpu
tiktoken
python-dotenv
beautifulsoup4
chromadb
pymongo
scikit-learn
unstructured
unstructured-inference 
unstructured_pytesseract
pypdf
pi-heif
pdf2image
pytesseract
pymupdf